+++ src/Blumchen/Subscriptions/SubscriptionOptionsBuilder.cs
@@ -129,23 +129,3 @@ public sealed class SubscriptionOptionsBuilder
 
     }
 }
-
-public class ObjectTracingConsumer: IHandler<object>
-{
-    private static ulong _counter = 0;
-    public Task Handle(object value)
-    {
-        Interlocked.Increment(ref _counter);
-        return Console.Out.WriteLineAsync();
-    }
-}
-
-public interface IErrorProcessor
-{
-    Func<Exception, Task> Process { get; }
-}
-
-public record ConsoleOutErrorProcessor: IErrorProcessor
-{
-    public Func<Exception, Task> Process => exception => Console.Out.WriteLineAsync($"record id:{0} resulted in error:{exception.Message}");
-}
